# chore(core): hot path optimisation in CSV parser