--- a/core/src/main/java/io/questdb/std/Files.java
+++ b/core/src/main/java/io/questdb/std/Files.java
@@ -128,6 +128,8 @@ public final class Files {
         return lpsz != null && exists0(lpsz.address());
     }
 
+    public static native long noop();
+
     public static void fadvise(long fd, long offset, long len, int advise) {
         if (Os.type == Os.LINUX_AMD64 || Os.type == Os.LINUX_ARM64) {
             fadvise0(fd, offset, len, advise);