… core/src/main/resources/io/questdb/bin/windows/libquestdb.dll