Another 20% perf improvement. Tested on windows and clickhouse file.

355s > 255s, disk monitor shows up to 340MB/s read, which is up from 240MB/s

on apple M1 performance gains are simply unreal. 156s for 76GB file with 16GB available RAM. 500MB/s