@ideoma 2022-10-11
#### [PR Coverage check]

:heart_eyes: **pass** : 17 / 17 (100.00%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/cutlass/text/DefaultTextConfiguration.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/cutlass/text/AbstractTextLexer.java|15|15|100.00%|